fix(platform-server): destroy PlatformRef when the render fails
Previously, the `PlatformRef` was destroyed only when the render succeeded.
But in case of a render error (e.g. due to a failed `APP_INITIALIZER`
or an error in a `constructor` of a root component), the `PlatformRef` was left
not destroyed. Not destroying `PlatformRef` means not calling `ngOnDestroy`
on services - which can lead to memory leaks when those services had allocated
resources to be released their `ngOnDestroy` hooks.

Now the `PlatformRef` is destroyed after every finished render, regardless
if it was successful or not, which helps to avoid memory leaks in applications